\name{hackR-package}
\alias{hackR-package}
\docType{package}
\title{Tools for Leet Hackers in Search for a New Handle}
\description{
  A set of tools for l33tifying crappy movie hacker handles, or
  even generating new ones.
}
\details{
\tabular{ll}{
  Package: \tab hackR\cr
  Type: \tab Package\cr
  License: \tab BSD 2-Clause\cr
}
}
\references{
  \url{https://github.com/wrathematics/hackR}
}
\author{
  Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
