% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackR-package.r
\docType{package}
\name{hackR-package}
\alias{hackR-package}
\title{Tools for Leet Hackers in Search for a New Handle}
\description{
A set of tools for l33tifying crappy movie hacker handles, or even
generating new ones.
}
\details{
\tabular{ll}{
  Package: \tab hackR \cr 
  Type: \tab Package \cr 
  License: \tab BSD 2-Clause \cr
}
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\references{
\url{https://github.com/wrathematics/hackR}
}
\keyword{Package}

