% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackr.R
\name{hackr_best_stories}
\alias{hackr_best_stories}
\title{Get Best Stories}
\usage{
hackr_best_stories(n = 10, unnest = FALSE)
}
\arguments{
\item{n}{The number of top stories to return, defaults to 10.}

\item{unnest}{Should "kids" column be nested? Defaults to TRUE.}
}
\description{
Fetches 10 best stories. You can fetch more by adjusting the n parameter. By
default, a tibble is returned and the "kids" column nested (if present).
}
