% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackr.R
\name{hackr_raw}
\alias{hackr_raw}
\title{Get Hacker News via Raw Endpoint}
\usage{
hackr_raw(x, ...)
}
\arguments{
\item{x}{Endpoint}

\item{...}{Any arguments you can pass to \code{jsonlite::fromJSON}}
}
\description{
Provide the raw endpoint, do not include the prefix
(https://hacker-news.firebaseio.com/v0), hackr will add it for you. The
forward slash must be included, i.e. /item/8863.json
}
