% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackr.R
\name{hackr_top_stories}
\alias{hackr_top_stories}
\title{Get Top Stories}
\usage{
hackr_top_stories(n = 10, unnest = FALSE)
}
\arguments{
\item{n}{The number of top stories to return, defaults to 10.}

\item{unnest}{Should "kids" column be nested? Defaults to TRUE.}
}
\description{
Fetches the top 10 stories. You can fetch more by adjusting the n parameter.
By default, a tibble is returned and the "kids" column nested (if present).
}
